Fix: IPC messages not reaching all terminals in multi-terminals
Due to the System V Message Queue structure, a message was deleted from the queue when it was read. This caused a notification sent when three or more terminals were open to be received only by the fastest terminal, preventing others from seeing the message (a race condition).

Changes:
- send_ipc_message: Instead of a single message, duplicate messages are sent based on the number of active instances in the system (flooding).
- ipc_listener: Terminals consuming their own sent messages are now allowed to wait (usleep) to give others a chance to read the message.

Thus, the broadcast logic was simulated manually.